Move ExtractedInput to scaffolding, fix compile errors
